reimplement #10142
