+++ core/src/com/unciv/logic/civilization/CivConstructions.kt
@@ -5,6 +5,7 @@ import com.unciv.logic.city.City
 import com.unciv.models.Counter
 import com.unciv.models.ruleset.Building
 import com.unciv.models.ruleset.INonPerpetualConstruction
+import com.unciv.models.ruleset.unique.StateForConditionals
 import com.unciv.models.ruleset.unique.UniqueType
 import com.unciv.models.ruleset.unit.BaseUnit
 import com.unciv.models.stats.Stat
@@ -64,6 +65,7 @@ class CivConstructions : IsPartOfGameInfoSerialization {
     fun tryAddFreeBuildings() {
         addFreeStatsBuildings()
         addFreeSpecificBuildings()
+        addFreeBuildings()
     }
 
     /** Common to [hasFreeBuildingByName] and [getFreeBuildingNames] - 'has' doesn't need the whole set, one enumeration is enough.
@@ -136,6 +138,30 @@ class CivConstructions : IsPartOfGameInfoSerialization {
             freeSpecificBuildingsProvided.addToMapOfSets(building.name, city.id)
             addFreeBuilding(city.id, building.name)
             city.cityConstructions.constructionComplete(building)
+        }
+    }
+
+    fun addFreeBuildings() {
+        val autoGrantedBuildings = civInfo.gameInfo.ruleset.buildings.values
+            .filter { it.hasUnique(UniqueType.GainBuildingWhereBuildable) }
+
+        // "Gain a free [buildingName] [cityFilter]"
+        val freeBuildingsFromCiv = civInfo.getMatchingUniques(UniqueType.GainFreeBuildings, StateForConditionals.IgnoreConditionals)
+        for (city in civInfo.cities) {
+            val freeBuildingsFromCity = city.getMatchingLocalOnlyUniques(UniqueType.GainFreeBuildings, StateForConditionals.IgnoreConditionals)
+            val freeBuildingUniques = (freeBuildingsFromCiv + freeBuildingsFromCity)
+                .filter { city.matchesFilter(it.params[1]) && it.conditionalsApply(StateForConditionals(city.civ, city)) }
+            for (unique in freeBuildingUniques){
+                val freeBuilding = city.civ.getEquivalentBuilding(unique.params[0])
+                city.cityConstructions.freeBuildingsProvidedFromThisCity.addToMapOfSets(city.id, freeBuilding.name)
+
+                if (city.cityConstructions.containsBuildingOrEquivalent(freeBuilding.name)) continue
+                city.cityConstructions.addBuilding(freeBuilding)
+            }
+
+            for (building in autoGrantedBuildings)
+                if (building.isBuildable(city.cityConstructions))
+                    city.cityConstructions.addBuilding(building)
         }
     }
